// types.ts
export interface ContactFormData {
	name: string
	email: string
	phone: string
	subject: string
	message: string
}
export interface RequestFormData {
	name: string
	phone: string
	email: string
	vehicleType: string
	location: string
	destination: string
	urgency: string
	subject: string
	message: string
	description: string
}